players_elo_df elo column is not getting u pdated
